debug

+++ library/signal/derive.ts
@@ -26,9 +26,11 @@ export function createDerive<T>(deriver: SignalDeriver<T>, staticDependencies?: 
 
 		activate = (set: SignalSetter<T>) => {
 			function update() {
-				signalSyncContextStack.push(new Set())
+				const _syncContext = new Set()
+				signalSyncContextStack.push()
 				const value = deriver()
 				const syncContext = signalSyncContextStack.pop()!
+				if (syncContext !== _syncContext) throw new Error("Now thats unexpected")
 				syncContext.delete(self)
 				const dependenciesToDelete = new Set<SignalReadable>()
 				for (const [dependency, subscription] of dependencyToSubscriptionMap.entries()) {
